+++ net/atm/mpc.h
@@ -1,6 +1,7 @@
 #ifndef _MPC_H_
 #define _MPC_H_
 
+#include <linux/types.h>
 #include <linux/atm.h>
 #include <linux/atmmpc.h>
 #include <linux/skbuff.h>
@@ -50,6 +51,6 @@ struct atm_mpoa_qos *atm_mpoa_search_qos(uint32_t dst_ip);
 int atm_mpoa_delete_qos(struct atm_mpoa_qos *qos);
 
 /* Display QoS entries. This is for the procfs */
-void atm_mpoa_disp_qos(char *page, int *len);
+void atm_mpoa_disp_qos(char *page, ssize_t *len);
 
 #endif /* _MPC_H_ */
